daemon: move base program compile to loader pkg
This package is responsible for the management and compilation of datapath
programs for Endpoints, so logically collocating the compilation of base
programs within said package makes sense from an organizational perspective in
the code. Along with this, move the analysis of requirements for running the
daemon into `pkg/datapath/linux`.

Signed-off by: Ian Vernon <ian@cilium.io>
