+++ daemon/policy.go
@@ -270,7 +270,7 @@ func (d *Daemon) policyAdd(sourceRules policyAPI.Rules, opts *policy.AddOptions,
 	if newPrefixLengths && !bpfIPCache.BackedByLPM() {
 		// Only recompile if configuration has changed.
 		logger.Debug("CIDR policy has changed; recompiling base programs")
-		if err := d.compileBase(); err != nil {
+		if err := d.Datapath().Loader().CompileBasePrograms(d.ctx, d, d.mtuConfig.GetDeviceMTU(), d.iptablesManager, d.l7Proxy, d.ipam); err != nil {
 			_ = d.prefixLengths.Delete(prefixes)
 			metrics.PolicyImportErrors.Inc()
 			err2 := fmt.Errorf("Unable to recompile base programs: %s", err)
@@ -576,7 +576,7 @@ func (d *Daemon) policyDelete(labels labels.LabelArray, res chan interface{}) {
 	if !bpfIPCache.BackedByLPM() && prefixesChanged {
 		// Only recompile if configuration has changed.
 		log.Debug("CIDR policy has changed; recompiling base programs")
-		if err := d.compileBase(); err != nil {
+		if err := d.Datapath().Loader().CompileBasePrograms(d.ctx, d, d.mtuConfig.GetDeviceMTU(), d.iptablesManager, d.l7Proxy, d.ipam); err != nil {
 			log.WithError(err).Error("Unable to recompile base programs")
 		}
 	}
