+++ pkg/datapath/loader/loader.go
@@ -56,6 +56,15 @@ type Loader struct {
 
 	// templateCache is the cache of pre-compiled datapaths.
 	templateCache *objectCache
+
+	canDisableDwarfRelocations bool
+}
+
+// NewLoader returns a new loader.
+func NewLoader(canDisableDwarfRelocations bool) *Loader {
+	return &Loader{
+		canDisableDwarfRelocations: canDisableDwarfRelocations,
+	}
 }
 
 // Init initializes the datapath cache with base program hashes derived from
